+++ internal/usecase/room.go
@@ -24,7 +24,7 @@ func NewRoomUseCase(store *postgres.Store) *RoomUseCase {
 	return &RoomUseCase{repo: store.RoomRepo}
 }
 
-func (uc *RoomUseCase) CreateRoom(ctx context.Context, name string, ownerID int) (*domain.Room, error) {
+func (uc *RoomUseCase) CreateRoom(ctx context.Context, name string, ownerID int64) (*domain.Room, error) {
 	code, err := generateInviteCode()
 	if err != nil {
 		return nil, err
@@ -42,7 +42,7 @@ func (uc *RoomUseCase) CreateRoom(ctx context.Context, name string, ownerID int)
 	return room, nil
 }
 
-func (uc *RoomUseCase) JoinRoom(ctx context.Context, userID int, roomID int, inviteCode string) (*domain.Room, error) {
+func (uc *RoomUseCase) JoinRoom(ctx context.Context, userID int64, roomID int64, inviteCode string) (*domain.Room, error) {
 	room, err := uc.repo.FindByID(ctx, roomID)
 	if err != nil {
 		return nil, err
@@ -56,7 +56,7 @@ func (uc *RoomUseCase) JoinRoom(ctx context.Context, userID int, roomID int, inv
 	return room, nil
 }
 
-func (uc *RoomUseCase) GetRoomsByUser(ctx context.Context, userID int) ([]domain.Room, error) {
+func (uc *RoomUseCase) GetRoomsByUser(ctx context.Context, userID int64) ([]domain.Room, error) {
 	return uc.repo.GetByUserID(ctx, userID)
 }
